% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{acyl_cation_mass}
\alias{acyl_cation_mass}
\title{Calculate Acyl Cation m/z}
\usage{
acyl_cation_mass(carbon, double_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{double_bond}{Integer. Double bond count}
}
\value{
Numeric. Calculated m/z
}
\description{
Calculates fatty acid/acyl cation m/z from chain composition
}
\keyword{internal}
