% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{acyl_chain_loss_mass}
\alias{acyl_chain_loss_mass}
\title{Calculate Neutral Loss m/z}
\usage{
acyl_chain_loss_mass(carbon, double_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{double_bond}{Integer. Double bond count}
}
\value{
Numeric. Calculated m/z
}
\description{
For acyl chain loss from precursor
}
\keyword{internal}
