% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{characterize_lipid_by_class}
\alias{characterize_lipid_by_class}
\title{Comprehensive Lipid MS/MS Characterization}
\usage{
characterize_lipid_by_class(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  sn_params,
  adduct,
  lipid_class
)
}
\arguments{
\item{ms_scan_prop}{List. MS scan properties}

\item{ms2_tolerance}{Numeric. MS2 tolerance}

\item{theoretical_mz}{Numeric. Precursor m/z}

\item{total_carbon}{Integer. Total carbons}

\item{total_double_bond}{Integer. Total double bonds}

\item{sn_params}{List. Parameters for sn-position estimation}

\item{adduct}{List. Adduct information}

\item{lipid_class}{Character. Lipid class code}
}
\value{
LipidMolecule list or NULL
}
\description{
Master function that dispatches to appropriate lipid class function
}
\keyword{internal}
