% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{count_fragment_existence}
\alias{count_fragment_existence}
\title{Count Fragment Existence in Spectrum}
\usage{
count_fragment_existence(spectrum, query_fragments, ms2_tolerance)
}
\arguments{
\item{spectrum}{Data frame. Query spectrum}

\item{query_fragments}{List of lists. Each with Mass and Intensity}

\item{ms2_tolerance}{Numeric. Mass tolerance}
}
\value{
List with count and average_intensity
}
\description{
Count Fragment Existence in Spectrum
}
\keyword{internal}
