% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{create_lipid_annotation_as_level1}
\alias{create_lipid_annotation_as_level1}
\title{Create Lipid Molecule Object (Level 1 - Class Only)}
\usage{
create_lipid_annotation_as_level1(
  lipid_header,
  lbm_class,
  total_carbon,
  total_double_bond,
  score = 0
)
}
\arguments{
\item{lipid_header}{Character. Lipid class abbreviation (e.g., "PC")}

\item{lbm_class}{Character. Lipid classification code}

\item{total_carbon}{Integer. Total carbon atoms}

\item{total_double_bond}{Integer. Total double bonds}

\item{score}{Numeric. Annotation score}
}
\value{
LipidMolecule list object
}
\description{
Create Lipid Molecule Object (Level 1 - Class Only)
}
\keyword{internal}
