% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{ether_bond_acyl_loss}
\alias{ether_bond_acyl_loss}
\title{Calculate Ether Bond Acyl Loss m/z}
\usage{
ether_bond_acyl_loss(carbon, double_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{double_bond}{Integer. Double bond count}
}
\value{
Numeric. Calculated m/z
}
\description{
For ether-linked acyl chains
}
\keyword{internal}
