% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_ceramide_molecule_obj_as_level2}
\alias{get_ceramide_molecule_obj_as_level2}
\title{Create Ceramide Level 2 (Sphingobase + Acyl)}
\usage{
get_ceramide_molecule_obj_as_level2(
  lipid_class,
  lbm_class,
  hydroxy_string,
  sph_carbon,
  sph_double,
  acyl_carbon,
  acyl_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character. Lipid class}

\item{lbm_class}{Character. Classification code}

\item{hydroxy_string}{Character. Hydroxylation pattern ("m", "d", "t")}

\item{sph_carbon}{Integer. Sphingobase carbons}

\item{sph_double}{Integer. Sphingobase double bonds}

\item{acyl_carbon}{Integer. Acyl chain carbons}

\item{acyl_double}{Integer. Acyl chain double bonds}

\item{score}{Numeric. Match score}
}
\value{
LipidMolecule list
}
\description{
Create Ceramide Level 2 (Sphingobase + Acyl)
}
\keyword{internal}
