% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidomics_core.R
\name{get_database_start_index}
\alias{get_database_start_index}
\title{Get database start index for m/z search}
\usage{
get_database_start_index(mz, tolerance, molecules)
}
\arguments{
\item{mz}{Numeric. Query m/z value}

\item{tolerance}{Numeric. Mass tolerance}

\item{molecules}{Data frame. Sorted reference molecules}
}
\value{
Integer. Starting index for m/z window search
}
\description{
Performs binary search to find starting position for m/z window filtering
}
\keyword{internal}
