% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_phospholipid_molecule_obj_as_level3}
\alias{get_phospholipid_molecule_obj_as_level3}
\title{Create Phospholipid Level 3 (sn-Position Specified)}
\usage{
get_phospholipid_molecule_obj_as_level3(
  lipid_class,
  lbm_class,
  sn1_carbon,
  sn1_double,
  sn2_carbon,
  sn2_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character. Lipid class}

\item{lbm_class}{Character. Classification code}

\item{sn1_carbon}{Integer. sn1 position carbons}

\item{sn1_double}{Integer. sn1 position double bonds}

\item{sn2_carbon}{Integer. sn2 position carbons}

\item{sn2_double}{Integer. sn2 position double bonds}

\item{score}{Numeric. Match score}
}
\value{
LipidMolecule list
}
\description{
Maintains sn-position order without reordering chains.
}
\keyword{internal}
