% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{is_peak_found_with_criterion}
\alias{is_peak_found_with_criterion}
\title{Find Peak Within m/z Range}
\usage{
is_peak_found_with_criterion(spectrum, begin_mz, end_mz, threshold)
}
\arguments{
\item{spectrum}{Data frame with Mass and Intensity columns}

\item{begin_mz}{Numeric. Lower m/z boundary}

\item{end_mz}{Numeric. Upper m/z boundary}

\item{threshold}{Numeric. Minimum intensity threshold}
}
\value{
Logical. TRUE if peak found in range
}
\description{
Find Peak Within m/z Range
}
\keyword{internal}
