% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{return_annotation_result}
\alias{return_annotation_result}
\title{Return Annotation Result}
\usage{
return_annotation_result(
  lipid_header,
  lbm_class,
  hydrogen_string,
  theoretical_mz,
  adduct,
  total_carbon,
  total_double_bond,
  total_oxidized,
  candidates = NULL,
  acyl_count_in_molecule = 1
)
}
\arguments{
\item{lipid_header}{Character. Lipid class header}

\item{lbm_class}{Character. Classification code}

\item{hydrogen_string}{Character. Hydroxylation notation}

\item{theoretical_mz}{Numeric. Theoretical m/z}

\item{adduct}{List. Adduct information (IonMode, AdductIonName, etc.)}

\item{total_carbon}{Integer. Total carbons}

\item{total_double_bond}{Integer. Total double bonds}

\item{total_oxidized}{Integer. Total oxidized positions}

\item{candidates}{List. Candidate molecules}

\item{acyl_count_in_molecule}{Integer. Expected acyl count}
}
\value{
LipidMolecule. Best candidate or level 1/2 annotation
}
\description{
Helper to return best candidate or create sublevel annotation
}
\keyword{internal}
